Move EAct behavior into SETabController and remove ExtraTabController
